pyproject.toml
[tool.poetry]
name = "graduation-project"
version = "0.1.0"
description = ""
authors = ["yaaa3ser <yasserissa1411@gmail.com>"]
readme = "README.md"
packages = [{include = "graduation_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
drf-yasg = "^1.21.5"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
